added get, create, delete comment endpoints

+++ src/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { PrismaService } from '../prisma/prisma.service';
 
@@ -7,8 +7,7 @@ export class PostsService {
     constructor(private prisma: PrismaService) {}
 
     async create(createPostDto: CreatePostDto, sub: number) {
-        const post = await this.prisma.post.create({ data: { ...createPostDto, authorId: sub } });
-        return post;
+        return await this.prisma.post.create({ data: { ...createPostDto, authorId: sub } });
     }
 
     async findAll() {
@@ -16,6 +15,11 @@ export class PostsService {
     }
 
     async remove(id: number) {
+        const post = await this.prisma.post.findUnique({ where: { id } });
+        if (!post) {
+            throw new NotFoundException({ message: 'Post not found', statusCode: 404, error: 'post' });
+        }
+
         return await this.prisma.post.delete({ where: { id } });
     }
 }
